moved uri validation to utility script
