+++ src/scripts/utility/dateUtil.ts
@@ -1,3 +1,3 @@
 export const getCurrentYear = (): number => {
-  return new Date().getFullYear()
-}
+  return new Date().getFullYear();
+};
