+++ src/scripts/utility/text.ts
@@ -1,0 +1,55 @@
+const validURIType: string[] = [
+  "acap://",
+  "app://",
+  "file://",
+  "ftp://",
+  "ftps://",
+  "http://",
+  "https://",
+  "icap://",
+  "imap://",
+  "irc://",
+  "mailto:",
+  "ms-access:",
+  "ms-excel:",
+  "ms-infopath:",
+  "ms-powerpoint:",
+  "ms-project:",
+  "ms-publisher:",
+  "ms-spd:",
+  "ms-visio:",
+  "ms-word:",
+  "msteams://",
+  "mtqp://",
+  "nntp://",
+  "pop://",
+  "rsync://",
+  "rtmp://",
+  "rtsp://",
+  "sftp://",
+  "shortcuts://",
+  "smb://",
+  "smtp://",
+  "ssh://",
+  "telnet://",
+  "viber://",
+  "webcal://",
+  "wss://",
+  "xmpp://",
+  "zoomus://",
+  "zoommtg://"
+];
+
+export const validateUrl = (originalUrl: string) => {
+  try {
+    if (
+      !validURIType.some((uriType) => originalUrl.toLowerCase().startsWith(uriType.toLowerCase()))
+    ) {
+      return false;
+    } else {
+      return true;
+    }
+  } catch (error) {
+    console.error("Error validating the url: ", error);
+  }
+};
